issue-38: Session class moved to Server and finished Kickplayer
